Export internal types required for portability from a separate entrypoint (#433)
