+++ libs/langgraph/src/_types.ts
@@ -1,0 +1,16 @@
+// Export types for portability, these shouldn't be needed/used directly
+export {
+  type AnyValue,
+  type WaitForNames,
+  type DynamicBarrierValue,
+  type LastValue,
+  type NamedBarrierValue,
+  type Topic,
+} from "./channels/index.js";
+
+export {
+  type AnnotationRoot,
+  type StateType,
+  type UpdateType,
+  type NodeType,
+} from "./graph/index.js";
